+++ apps/web/app/components/AddBlockPopover/types.ts
@@ -1,8 +1,16 @@
+import type { BlockListCategory, BlockTemplateVariation } from '~/composables/useBlocksList/types';
+
 export interface PopoverPosition {
   left: number;
   top: number;
   opacity: number;
   arrowLeft: number;
   arrowTop: number;
-  arrowDir: 'up' | 'down';
+  arrowDirection: 'up' | 'down';
+}
+
+export interface FlatVariation {
+  category: BlockListCategory;
+  variation: BlockTemplateVariation;
+  idx: number;
 }
