--- apps/web/app/composables/useAddBlockPopoverPanel/index.ts
@@ -1,2 +1,0 @@
-export * from './useAddBlockPopoverPanel';
-export * from './types';
